pyproject.toml
[tool.poetry]
name = "llm-web-application"
version = "0.1.0"
description = ""
authors = ["Serhii Chukin <chukin2003.04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
openai = "^1.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
